requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=10.0.0
scikit-learn>=1.3.0
seaborn>=0.12.0
pandas>=2.0.0
tqdm>=4.65.0
icrawler>=0.6.6
jupyter>=1.0.0
ipykernel>=6.25.0
gdown>=4.7.0
